Move exceptions and references out of amati.__init__ to avoid polluting the namespace with imports irrelevant to amati
